fix: update get HTTP request for mapping recipe ingredients
