--- src/pages/hikings/eatings/eatingContract.ts
@@ -1,31 +1,0 @@
-export interface IEating {
-  id: string;
-  dayNumber: number;
-  eatingTimeId: string;
-  eatingTime: {
-    id: string;
-    name: string;
-  };
-  recipeId: string;
-  recipe: {
-    name: string;
-    kkal: string;
-  };
-}
-export interface IGetEatingsPayload {
-  page: number;
-}
-
-export interface IEatingCreatePayload {
-  dayNumber: number;
-  eatingTimeId: string;
-  eatingTime: {
-    id: string;
-    name: string;
-  };
-  recipeId: string;
-  recipe: {
-    name: string;
-    kkal: string;
-  };
-}
